Fix login function with --continue-on-success
The login function does not return True or False in any case. 

This leads to a problem if the flag --continue-on-success is used in combination with modules.
The login function does not return anything explicit, but it is has to return true in order to launch a module on a successful login.

Signed-off-by: trietend <212042+trietend@users.noreply.github.com>